\name{cycle}
\alias{cycle}
\title{Cyclically rotatation of elements of a vector}
\description{This function Cyclically rotate elements of a given vector}
\usage{cycle(x)}
\arguments{
	\item{x}{A vector}
	}
\value{cyclically rotated vector}
\author{Baidya Nath Mandal <mandal.stat@gmail.com>}
\keyword{internal}