\name{dpf}
\alias{dpf}
\title{Desired treatment versus position incidence matrix}
\description{This function returns desired treatment versus position incidence matrix for a proper block design with parameters v, b, r, k to be (nearly) position balanced}
\usage{dpf(v, b, r, k)}
\arguments{
	\item{v}{Number of treatments}
	\item{b}{Number of blocks}
	\item{r}{Number of replications of each treatment}
	\item{k}{Block size}
	}
\value{Desired treatment versus position incidence matrix}
\author{Baidya Nath Mandal <mandal.stat@gmail.com>}
\keyword{internal}