\name{psfreq}
\alias{psfreq}
\title{Treatment versus Position incidence matrix}
\description{This function returns Treatment versus Position incidence matrix of a given proper block design}
\usage{psfreq(design)}
\arguments{
	\item{design}{Block design specified in the form of a b x k matrix with elements labelled as 1 to v whre b is number of blocks, k is block size and v is number of treatments}
	}
\value{Treatment versus Position incidence matrix}
\author{Baidya Nath Mandal <mandal.stat@gmail.com>}
\keyword{internal}